map refactor
