+++ app/map/map.tsx
@@ -1,167 +1,96 @@
 "use client";
-import { useRef, useState, useEffect } from "react";
+import { useRef, useState, useMemo, useCallback } from "react";
 
-import mapboxgl from "mapbox-gl";
+import type { MapRef, GeoJSONSource } from "react-map-gl";
+import {
+  Map,
+  Source,
+  Layer,
+  Popup,
+  GeolocateControl,
+  FullscreenControl,
+  NavigationControl,
+  ScaleControl,
+} from "react-map-gl";
 
-import ErrorComponent from "./error";
+import { placesLayer, clusterLayer } from "./layers";
 
-export default function Map({ Places }: any) {
-  mapboxgl.accessToken = process.env.NEXT_PUBLIC_MAPBOX_TOKEN || "";
+const MAPBOX_TOKEN = process.env.NEXT_PUBLIC_MAPBOX_TOKEN; // Set your mapbox token here
 
-  const map = useRef<mapboxgl.Map | any>(null);
-  const mapContainer = useRef<any>(null);
-  const [clickedPlace, setClickedPlace] = useState<any>(null);
-  const [clickedClusterIds, setClickedClusterIds] = useState<number[]>([]);
-  const [lng, setLng] = useState<number>(-70.6109);
-  const [lat, setLat] = useState<number>(-33.4983);
-  const [zoom, setZoom] = useState<number>(16);
-  const [error, setError] = useState<Error>();
+export default function ReactMap(Places: any) {
+  const mapRef = useRef<MapRef>(null);
+  const [hoverInfo, setHoverInfo] = useState<any>(null);
 
-  useEffect(() => {
-    if (map.current) return;
-    map.current = new mapboxgl.Map({
-      container: mapContainer.current,
-      style: "mapbox://styles/mapbox/dark-v11",
-      bounds: [
-        [-70.6162, -33.5018],
-        [-70.6054, -33.4955],
-      ],
+  const onHover = useCallback((event: any) => {
+    const place = event.features && event.features[0];
+    setHoverInfo({
+      longitude: event.lngLat.lng,
+      latitude: event.lngLat.lat,
+      place: place ? place.properties.name : null,
     });
+  }, []);
 
-    map.current.on("move", () => {
-      if (map.current) {
-        setLng(map.current.getCenter().lng.toFixed(4));
-        setLat(map.current.getCenter().lat.toFixed(4));
-        setZoom(map.current.getZoom().toFixed(2));
-      } else {
-        setError(new Error("Map is not initialized"));
+  const selectedPlace = (hoverInfo && hoverInfo.place) || null;
+  const filter = useMemo(() => ["in", "name", selectedPlace], [selectedPlace]);
+
+  const onClick = (event: any) => {
+    const feature = event.features[0];
+    const clusterId = feature ? feature.properties.cluster_id : null;
+
+    const mapboxSource = mapRef.current.getSource("places") as GeoJSONSource;
+
+    mapboxSource.getClusterExpansionZoom(clusterId, (err: any, zoom: any) => {
+      if (err) {
+        return;
       }
-    });
 
-    map.current.on("style.load", () => {
-      addLayers();
-    });
-
-    map.current.on("click", "places-circle", (places: any) => {
-      const [selectedPlace] = places.features;
-
-      console.log("selectedPlace clicked", selectedPlace);
-      setClickedPlace(selectedPlace.properties);
-    });
-
-    map.current.on("click", "cluster-circle", (clusters: any) => {
-      const [selectedCluster] = clusters.features;
-
-      console.log("selectedCluster clicked", selectedCluster);
-      setClickedClusterIds(selectedCluster.properties.ids.split(","));
-    });
-
-    const popupHover = new mapboxgl.Popup({
-      closeButton: false,
-      closeOnClick: false,
-      className: "text-black",
-    });
-
-    map.current.on("mouseenter", "places-circle", (places: any) => {
-      const [place] = places.features;
-
-      const coordinates = place.geometry.coordinates.slice();
-      const identifier = place.properties.identifier;
-
-      map.current.getCanvas().style.cursor = "pointer";
-      popupHover.setLngLat(coordinates).setHTML(identifier).addTo(map.current);
-    });
-
-    map.current.on("mouseleave", "places-circle", () => {
-      map.current.getCanvas().style.cursor = "";
-      popupHover.remove();
-    });
-
-    function addLayers() {
-      map.current.addSource("places", {
-        type: "geojson",
-        data: Places,
-        cluster: true,
-        clusterProperties: { ids: ["concat", ["concat", ["get", "identifier"], ","]] },
-        clusterRadius: 10,
+      mapRef.current.easeTo({
+        center: feature.geometry.coordinates,
+        zoom,
+        duration: 500,
       });
-
-      map.current.addLayer({
-        id: "places-circle",
-        type: "circle",
-        filter: ["!", ["has", "point_count"]],
-        source: "places",
-        paint: {
-          "circle-color": [
-            "match",
-            ["at", 0, ["get", "categories"]],
-            "classroom",
-            "#FF8C00",
-            "shop",
-            "#0ef305",
-            "other",
-            "#e55e5e",
-            "#ccc",
-          ],
-          "circle-radius": 10,
-        },
-      });
-
-      map.current.addLayer({
-        id: "cluster-circle",
-        type: "circle",
-        filter: ["has", "point_count"],
-        source: "places",
-        paint: {
-          "circle-color": "#FF0000",
-          "circle-radius": 15,
-        },
-      });
-    }
-  }, [Places]);
-
-  if (error) return <ErrorComponent error={error} reset={() => setError(undefined)} />;
+    });
+  };
 
   return (
-    <div className="h-full w-full relative">
-      <div className="z-10 absolute top-0 left-0 flex-col">
-        <div className="bg-sidebar-color py-1.5 px-3 z-10  m-3 rounded-s">
-          Longitude: {+lng} | Latitude: {+lat} | Zoom: {+zoom}
-        </div>
-        <div className="bg-sidebar-color py-1.5 px-3 z-10  m-3 rounded-s flex-row">
-          <button
-            onClick={() => map.current.setStyle("mapbox://styles/mapbox/streets-v11")}
-            className="bg-white text-black rounded-s mx-2 py-1.5 mr-2"
+    <>
+      <Map
+        initialViewState={{
+          longitude: -70.6109,
+          latitude: -33.4983,
+          zoom: 16,
+        }}
+        mapStyle="mapbox://styles/mapbox/dark-v9"
+        mapboxAccessToken={MAPBOX_TOKEN}
+        interactiveLayerIds={[placesLayer.id, clusterLayer.id]}
+        // onClick={onClick}
+        onMouseMove={onHover}
+        ref={mapRef}
+        className="h-full w-full"
+        minZoom={10}
+      >
+        {/* {pins} */}
+        <GeolocateControl position="top-left" />
+        <FullscreenControl position="top-left" />
+        <NavigationControl position="top-left" />
+        <ScaleControl />
+        <Source id="places" type="geojson" data={Places.Places} cluster={true} clusterRadius={10}>
+          <Layer {...placesLayer} />
+          <Layer {...clusterLayer} />
+        </Source>
+        {selectedPlace ? (
+          <Popup
+            longitude={hoverInfo.longitude}
+            latitude={hoverInfo.latitude}
+            closeButton={false}
+            closeOnClick={false}
+            className="place"
+            offset={[0, -10]}
           >
-            Streets
-          </button>
-          <button
-            onClick={() => map.current.setStyle("mapbox://styles/mapbox/dark-v10")}
-            className="bg-white text-black rounded-s mx-2 py-1.5 mr-2"
-          >
-            Dark
-          </button>
-        </div>
-
-        {clickedPlace ? (
-          <div className="bg-sidebar-color py-1.5 px-3 z-10  m-3 rounded-s">
-            Id: {clickedPlace.identifier} | Nombre: {clickedPlace.name} | Categoria: {clickedPlace.categories} |
-            Descripcion: {clickedPlace.information}
-          </div>
+            {selectedPlace}
+          </Popup>
         ) : null}
-
-        {clickedClusterIds && clickedClusterIds.length > 0 ? (
-          <div className="bg-sidebar-color py-1.5 px-3 z-10  m-3 rounded-s flex-row flex">
-            {clickedClusterIds.map((clusterId) => (
-              <div key={clusterId} className="px-3">
-                {clusterId}
-              </div>
-            ))}
-          </div>
-        ) : null}
-      </div>
-
-      <div ref={mapContainer} className="h-full w-full" />
-    </div>
+      </Map>
+    </>
   );
 }
