--- app/map2/page.tsx
@@ -1,15 +1,0 @@
-import PlacesJSON from "../../data/places.json";
-
-// import Map from "./map";
-import MapComponent from "./map2";
-
-export default async function Page() {
-  return (
-    <>
-      {/* <Map Places={PlacesJSON} /> */}
-      <div className="h-full w-full">
-        <MapComponent Places={PlacesJSON} />
-      </div>
-    </>
-  );
-}
